Stop removing "\n" from list of characters that trigger indentation in electric-indent-mode.
Even though I intended this to only affect emacs-lisp-mode buffers, it
ended up messing with indentation behavior in other modes as
well (probably because the value of electric-indent-chars is not
buffer-local).

I *might* try to come up with a more refined version of the code that
this commit removes in the future. However, since I've been using
;-style comments less and less in my init-file the number of cases where
I would like Emacs to skip indentation has decreased significantly. If
this trend continues, I won't be needing specialized indentation
behavior in emacs-lisp-mode at all.